vcai-be/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pyautogen==0.4.0
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.0
celery==5.4.0
httpx==0.28.1
pydantic==2.10.2
pydantic-settings==2.6.1
aiofiles==24.1.0
websockets==14.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
psutil==6.1.0
